move UtilityFuncCRRA to interpolation
to allow inheritance from MetricObject
