+++ HARK/interpolation.py
@@ -12,8 +12,18 @@ from copy import deepcopy
 import numpy as np
 from scipy.interpolate import CubicHermiteSpline
 
-from HARK.utilities import CRRAutility, CRRAutilityP, CRRAutilityPP
 from HARK.core import MetricObject
+from HARK.utilities import (
+    CRRAutility,
+    CRRAutility_inv,
+    CRRAutility_invP,
+    CRRAutilityP,
+    CRRAutilityP_inv,
+    CRRAutilityP_invP,
+    CRRAutilityPP,
+    CRRAutilityPPP,
+    CRRAutilityPPPP,
+)
 
 
 def _isscalar(x):
@@ -4241,7 +4251,7 @@ class Curvilinear2DInterp(HARKinterpolator2D):
         zeta = a - x + c * tau
         eta = b + c * mu + d * tau
         theta = d * mu
-        alpha = (-eta + polarity * np.sqrt(eta ** 2.0 - 4.0 * zeta * theta)) / (
+        alpha = (-eta + polarity * np.sqrt(eta**2.0 - 4.0 * zeta * theta)) / (
             2.0 * theta
         )
         beta = mu * alpha + tau
@@ -4526,6 +4536,129 @@ def calc_log_sum(Vals, sigma):
 ###############################################################################
 
 
+class UtilityFuncCRRA(MetricObject):
+    """
+    A class for representing a CRRA utility function.
+
+    Parameters
+    ----------
+    CRRA : float
+        The coefficient of constant relative risk aversion.
+    """
+
+    distance_criteria = ["CRRA"]
+
+    def __init__(self, CRRA):
+        self.CRRA = CRRA
+
+    def __call__(self, c, order=0):
+        """
+        Evaluate the utility function at a given level of consumption c.
+
+        Parameters
+        ----------
+        c : float or np.ndarray
+            Consumption level(s).
+        order : int, optional
+            Order of derivative. For example, `order == 1` returns the
+            first derivative of utility of consumption, and so on. By default 0.
+
+        Returns
+        -------
+        float or np.ndarray
+            Utility (or its derivative) evaluated at given consumption level(s).
+        """
+        if order == 0:
+            return CRRAutility(c, self.CRRA)
+        else:  # order >= 1
+            return self.derivative(c, order)
+
+    def derivative(self, c, order=1):
+        """
+        The derivative of the utility function at a given level of consumption c.
+
+        Parameters
+        ----------
+        c : float or np.ndarray
+            Consumption level(s).
+        order : int, optional
+            Order of derivative. For example, `order == 1` returns the
+            first derivative of utility of consumption, and so on. By default 1.
+
+        Returns
+        -------
+        float or np.ndarray
+            Derivative of CRRA utility evaluated at given consumption level(s).
+
+        Raises
+        ------
+        ValueError
+            Derivative of order higher than 4 is not supported.
+        """
+        if order == 1:
+            return CRRAutilityP(c, self.CRRA)
+        elif order == 2:
+            return CRRAutilityPP(c, self.CRRA)
+        elif order == 3:
+            return CRRAutilityPPP(c, self.CRRA)
+        elif order == 4:
+            return CRRAutilityPPPP(c, self.CRRA)
+        else:
+            raise ValueError("Derivative of order {} not supported".format(order))
+
+    def inverse(self, u, order=(0, 0)):
+        """
+        The inverse of the utility function at a given level of utility u.
+
+        Parameters
+        ----------
+        u : float or np.ndarray
+            Utility level(s).
+        order : tuple, optional
+            Order of derivatives. For example, `order == (1,1)` represents
+            the first derivative of utility, inversed, and then differenciated
+            once. By default (0, 0), which is just the inverse of utility.
+
+        Returns
+        -------
+        float or np.ndarray
+            Inverse of CRRA utility evaluated at given utility level(s).
+
+        Raises
+        ------
+        ValueError
+            Higher order derivatives are not supported.
+        """
+        if order == (0, 0):
+            return CRRAutility_inv(u, self.CRRA)
+        elif order == (1, 0):
+            return CRRAutilityP_inv(u, self.CRRA)
+        elif order == (0, 1):
+            return CRRAutility_invP(u, self.CRRA)
+        elif order == (1, 1):
+            return CRRAutilityP_invP(u, self.CRRA)
+        else:
+            raise ValueError("Inverse of order {} not supported".format(order))
+
+    def der(self, c, order=1):
+        """
+        Short alias for derivative. See `self.derivative`.
+        """
+        return self.derivative(c, order)
+
+    def inv(self, c, order=(0, 0)):
+        """
+        Short alias for inverse. See `self.inverse`.
+        """
+        return self.inverse(c, order)
+
+    def derinv(self, u, order=(1, 0)):
+        """
+        Short alias for inverse. See `self.inverse`.
+        """
+        return self.inverse(u, order)
+
+
 class ValueFuncCRRA(MetricObject):
     """
     A class for representing a value function.  The underlying interpolation is
@@ -4727,7 +4860,7 @@ def main():
     if False:
 
         def f(x, y):
-            return 3.0 * x ** 2.0 + x * y + 4.0 * y ** 2.0
+            return 3.0 * x**2.0 + x * y + 4.0 * y**2.0
 
         def dfdx(x, y):
             return 6.0 * x + y
@@ -4779,10 +4912,10 @@ def main():
 
     if False:
         f = (
-            lambda x, y, z: 3.0 * x ** 2.0
+            lambda x, y, z: 3.0 * x**2.0
             + x * y
-            + 4.0 * y ** 2.0
-            - 5 * z ** 2.0
+            + 4.0 * y**2.0
+            - 5 * z**2.0
             + 1.5 * x * z
         )
 
@@ -4915,7 +5048,7 @@ def main():
     if False:
 
         def f(x, y):
-            return 3.0 * x ** 2.0 + x * y + 4.0 * y ** 2.0
+            return 3.0 * x**2.0 + x * y + 4.0 * y**2.0
 
         def dfdx(x, y):
             return 6.0 * x + y
@@ -4937,10 +5070,10 @@ def main():
 
     if False:
         f = (
-            lambda x, y, z: 3.0 * x ** 2.0
+            lambda x, y, z: 3.0 * x**2.0
             + x * y
-            + 4.0 * y ** 2.0
-            - 5 * z ** 2.0
+            + 4.0 * y**2.0
+            - 5 * z**2.0
             + 1.5 * x * z
         )
 
@@ -5034,7 +5167,7 @@ def main():
     if False:
 
         def f(x, y):
-            return 3.0 * x ** 2.0 + x * y + 4.0 * y ** 2.0
+            return 3.0 * x**2.0 + x * y + 4.0 * y**2.0
 
         def dfdx(x, y):
             return 6.0 * x + y
@@ -5069,10 +5202,10 @@ def main():
 
     if False:
         f = (
-            lambda x, y, z: 3.0 * x ** 2.0
+            lambda x, y, z: 3.0 * x**2.0
             + x * y
-            + 4.0 * y ** 2.0
-            - 5 * z ** 2.0
+            + 4.0 * y**2.0
+            - 5 * z**2.0
             + 1.5 * x * z
         )
